+ added checking user already authentication + added workManager + check coin price change logic impl. + added sending notification feature + changed app icon + create notification icon + some bug fixes

+++ app/src/main/java/com/example/cryptocurrencytracker/domain/model/CoinItem.kt
@@ -7,7 +7,7 @@ data class CoinItem(
     val name: String = "",
     val symbol: String = "",
     val image: String = "",
-    val current_price: Int = 0
+    val current_price: Double = 0.0
 ){
     fun toEntity(): CoinEntity{
         return CoinEntity(
