+++ app/src/main/java/com/example/cryptocurrencytracker/domain/model/CurrentPrice.kt
@@ -1,5 +1,5 @@
 package com.example.cryptocurrencytracker.domain.model
 
 data class CurrentPrice(
-    val usd: Int
+    val usd: Double
 )
